+++ .env.example
@@ -1,7 +1,6 @@
 AGENT__LOG_FORMAT=text
 AGENT__EVENT_STORE__TYPE=postgres
-AGENT__EVENT_STORE__CONNECTION_STRING="postgresql://demo_user:demo_pass@localhost:5432/demo"
-AGENT__URL="http://192.168.1.234:3033"
+AGENT__EVENT_STORE__CONNECTION_STRING="postgresql://demo_user:demo_pass@cqrs-postgres-db:5432/demo"
 
 AGENT__SECRET_MANAGER__STRONGHOLD_PATH="agent_secret_manager/tests/res/test.stronghold"
 AGENT__SECRET_MANAGER__STRONGHOLD_PASSWORD="secure_password"
