+++ agent_event_publisher_http/Cargo.toml
@@ -20,7 +20,6 @@ rustls = { version = "0.23", default-features = false, features = [
 ] }
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
 serde.workspace = true
-serde_json.workspace = true
 serde_with.workspace = true
 serde_yaml.workspace = true
 tokio.workspace = true
