+++ agent_event_publisher_http/README.md
@@ -2,22 +2,21 @@
 
 A simple HTTP event publisher for the SSI Agent.
 
-To make use of this publisher you need to configure it by creating a `config.yaml` file in this same directory. For each
-aggregate you want to publish events for, you need to set the following properties:
+To make use of this publisher you need to configure it by adding the `http` object to your configuration file.
 
 - The `target_url` is the URL to which the events will be published.
 - The `target_events` is a list of events that will be published to the `target_url`.
 
-Example `config.yaml`:
+Example:
 
 ```yaml
-target_url: &target_url "https://my-domain.example.org/ssi-event-subscriber"
-
-connection:
-  {
-    target_url: *target_url,
-    target_events: [SIOPv2AuthorizationResponseVerified],
-  }
+event_publishers:
+  http:
+    enabled: false
+    target_url: "https://my-domain.example.org/event-subscriber"
+    events:
+      server_config: []
+      credential: [UnsignedCredentialCreated, CredentialSigned]
 ```
 
 ### Available events
