+++ agent_secret_manager/src/lib.rs
@@ -24,6 +24,5 @@ pub async fn secret_manager() -> SecretManager {
                 .unwrap()
         }
         (snapshot_path, password, None, _, _) => SecretManager::generate(snapshot_path, password).await.unwrap(),
-        // _ => panic!("Unable to load or generate `SecretManager`. Please make sure to set both `AGENT__SECRET_MANAGER__STRONGHOLD_PATH` and `AGENT__SECRET_MANAGER__STRONGHOLD_PASSWORD` environment variables."),
     }
 }
