+++ agent_shared/src/lib.rs
@@ -4,7 +4,6 @@ pub mod domain_linkage;
 pub mod error;
 pub mod generic_query;
 pub mod handlers;
-pub mod issuance;
 pub mod url_utils;
 
 pub use ::config::ConfigError;
