+++ agent_store/Cargo.toml
@@ -9,11 +9,9 @@ agent_issuance = { path = "../agent_issuance" }
 agent_shared = { path = "../agent_shared" }
 agent_verification = { path = "../agent_verification" }
 
+async-trait = "0.1"
 cqrs-es.workspace = true
 postgres-es = "0.4.10"
-
-async-trait = "0.1"
-serde.workspace = true
 serde_json.workspace = true
 sqlx = { version = "0.7", features = [
     "postgres",
